Convert JS modules with .d.ts sidecars to TypeScript
Convert apiService.js, characterService.js, and data/rules.js to native
TypeScript and delete their hand-written .d.ts sidecars.

- Type exported constants/interfaces in rules.ts
- Export canonical CharacterData type from characterService.ts
- Type apiService.ts (cleanSchemaForApi, generateCharacterViaGemini)
- Update all import paths to drop .js extensions
- Remove @ts-expect-error comments in character.ts and import sites
- Fix latent type errors surfaced by proper typing (null/undefined guards)
