+++ src/components/sheet/SavingThrows.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { useCharacterStore } from '@/stores/character'
-import * as DND_RULES from '@/data/rules.js'
-import { formatMod } from '@/services/characterService.js'
+import * as DND_RULES from '@/data/rules'
+import { formatMod } from '@/services/characterService'
 
 const store = useCharacterStore()
 </script>
@@ -19,7 +19,7 @@ const store = useCharacterStore()
         <strong>
           {{
             formatMod(
-              store.abilityMods[key] +
+              (store.abilityMods[key] ?? 0) +
                 (store.currentCharacterData.proficiencies.savingThrows.includes(key)
                   ? store.profBonus
                   : 0),
