--- src/services/apiService.js.d.ts
@@ -1,3 +1,0 @@
-declare module '../services/apiService.js' {
-  export function generateCharacterViaGemini(userPrompt: string, schema: object): Promise<object>
-}
